pyproject.toml
[tool.poetry]
name = "cumplo-herald"
version = "0.1.1"
description = "A multi-channel notification service designed to keep Cumplo investors ahead with real-time alerts"
authors = ["Cristobal Sfeir <cnsfeir@uc.cl>"]
packages = [{ include = "cumplo_herald" }]

[tool.poetry.dependencies]
python = "^3.11"
arrow = "^1.2.3"
pydantic = "^2.1.1"
python-dotenv = "^1.0.0"
google-cloud-logging = "^3.5.0"
starlette = "<0.28.0"
fastapi = "^0.101.0"
uvicorn = "^0.23.1"
case-converter = "^1.1.0"
cumplo-common = { version = "^0.2.3", source = "cumplo-pypi" }

[tool.poetry.dev-dependencies]
pylint = "^2.17.2"
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.2.0"

[[tool.poetry.source]]
name = "cumplo-pypi"
url = "https://us-central1-python.pkg.dev/cumplo-scraper/cumplo-pypi/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
